refactor: Entity 패키지 변경
