+++ src/main/java/com/mobble/mobbleserver/domain/like/articleLike/ArticleLike.java
@@ -1,9 +1,9 @@
-package com.mobble.mobbleserver.refactor.like.articleLike.entity;
+package com.mobble.mobbleserver.domain.like.articleLike;
 
+import com.mobble.mobbleserver.domain.like.baseLike.BaseLike;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.like.baseLike.entity.BaseLike;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
