+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/BaseLike.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.like.baseLike.entity;
+package com.mobble.mobbleserver.domain.like.baseLike;
 
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
