+++ src/main/java/com/mobble/mobbleserver/domain/like/clubLike/ClubLike.java
@@ -1,9 +1,9 @@
-package com.mobble.mobbleserver.refactor.like.clubLike.entity;
+package com.mobble.mobbleserver.domain.like.clubLike;
 
+import com.mobble.mobbleserver.domain.like.baseLike.BaseLike;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
-import com.mobble.mobbleserver.refactor.like.baseLike.entity.BaseLike;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
