+++ src/main/java/com/mobble/mobbleserver/refactor/like/baseLike/service/LikeQueryService.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.like.baseLike.service;
 
 import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
-import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
+import com.mobble.mobbleserver.domain.like.baseLike.LikeType;
 
 public interface LikeQueryService {
