+++ src/test/java/com/mobble/mobbleserver/refactor/like/baseLike/entity/LikeTestFixture.java
@@ -1,8 +1,9 @@
 package com.mobble.mobbleserver.refactor.like.baseLike.entity;
 
+import com.mobble.mobbleserver.domain.like.baseLike.BaseLike;
 import com.mobble.mobbleserver.domain.member.Member;
 
-public class LikeTestFixture extends BaseLike{
+public class LikeTestFixture extends BaseLike {
 
     public static LikeTestFixture create(Member member) {
         LikeTestFixture like = new LikeTestFixture();
