+++ src/test/java/com/mobble/mobbleserver/refactor/like/clubLike/entity/ClubLikeTest.java
@@ -1,5 +1,6 @@
 package com.mobble.mobbleserver.refactor.like.clubLike.entity;
 
+import com.mobble.mobbleserver.domain.like.clubLike.ClubLike;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
